Add: 가까운 방향으로 회전하면서 화살표로 가리키는 함수 작성
