+++ public/js/convertData.js
@@ -12,7 +12,7 @@ const hexArr2decArr = hexArr => hexArr.map(hex => parseInt(hex, 16));
 
 const decArr2str = decArr => String.fromCharCode(...decArr);
 
-export default {
+export {
     receiveStr,
     str2arr,
     charArr2decArr,
